Refactor: move importData class files into an importData directory under the model directory
Some Util methods had to be changed from protected to public with this change.
