+++ src/main/java/edu/tamu/di/SAFCreator/model/importData/ImportDataProcessor.java
@@ -1,4 +1,4 @@
-package edu.tamu.di.SAFCreator;
+package edu.tamu.di.SAFCreator.model.importData;
 
 import javax.swing.JTextArea;
